refactored code for Drug and Study data ingestion
