--- CuDfDomino_Env.md
@@ -1,142 +1,0 @@
-Package              Version                
--------------------- -----------------------
-aiohttp              3.6.2                  
-appdirs              1.4.3                  
-async-timeout        3.0.1                  
-attrs                19.3.0                 
-backcall             0.1.0                  
-bleach               3.1.4                  
-bokeh                1.4.0                  
-certifi              2020.4.5.1             
-cffi                 1.14.0                 
-chardet              3.0.4                  
-click                7.1.1                  
-click-plugins        1.1.1                  
-cligj                0.5.0                  
-cloudpickle          1.3.0                  
-colorcet             2.0.1                  
-cryptography         2.8                    
-cudf                 0.13.0+0.ga2804c3.dirty
-cugraph              0.13.0+0.gac36e8c.dirty
-cuml                 0.13.0+0.g7544c43.dirty
-cupy                 7.3.0                  
-cusignal             0+unknown              
-cuspatial            0.13.0+0.g12ee694.dirty
-cuxfilter            0.13.0+0.g6e29cd2.dirty
-cycler               0.10.0                 
-cytoolz              0.10.1                 
-dask                 2.14.0                 
-dask-cuda            0.13.0                 
-dask-cudf            0.13.0+0.ga2804c3.dirty
-dask-xgboost         0.1.5                  
-datashader           0.10.0                 
-datashape            0.5.4                  
-decorator            4.4.2                  
-defusedxml           0.6.0                  
-distributed          2.14.0                 
-entrypoints          0.3                    
-fastavro             0.23.1                 
-fastrlock            0.4                    
-Fiona                1.8.9.post2            
-fsspec               0.6.3                  
-GDAL                 2.4.4                  
-geopandas            0.7.0                  
-HeapDict             1.0.1                  
-idna                 2.9                    
-imageio              2.8.0                  
-importlib-metadata   1.6.0                  
-ipykernel            5.2.0                  
-ipython              7.13.0                 
-ipython-genutils     0.2.0                  
-jedi                 0.16.0                 
-Jinja2               2.11.1                 
-joblib               0.14.1                 
-jsonschema           3.2.0                  
-jupyter-client       6.1.2                  
-jupyter-core         4.6.3                  
-jupyter-server-proxy 1.3.2                  
-kiwisolver           1.2.0                  
-llvmlite             0.31.0                 
-locket               0.2.0                  
-Markdown             3.2.1                  
-MarkupSafe           1.1.1                  
-matplotlib           3.2.1                  
-mistune              0.8.4                  
-msgpack              1.0.0                  
-multidict            4.7.5                  
-multipledispatch     0.6.0                  
-munch                2.5.0                  
-nbconvert            5.6.1                  
-nbformat             5.0.4                  
-neo4j                1.7.6                  
-neobolt              1.7.17                 
-neotime              1.7.4                  
-networkx             2.4                    
-notebook             6.0.3                  
-numba                0.48.0                 
-numpy                1.18.1                 
-nvstrings-cuda100    0.13.0                 
-olefile              0.46                   
-packaging            20.1                   
-pandas               0.25.3                 
-pandocfilters        1.4.2                  
-panel                0.6.4                  
-param                1.9.3                  
-parso                0.6.2                  
-partd                1.1.0                  
-pexpect              4.8.0                  
-pickleshare          0.7.5                  
-Pillow               7.1.1                  
-pip                  20.0.2                 
-prometheus-client    0.7.1                  
-prompt-toolkit       3.0.5                  
-psutil               5.7.0                  
-ptyprocess           0.6.0                  
-pyarrow              0.15.0                 
-pycparser            2.20                   
-pyct                 0.4.6                  
-pyee                 0.0.0                  
-Pygments             2.6.1                  
-pynvml               8.0.4                  
-pyOpenSSL            19.1.0                 
-pyparsing            2.4.7                  
-pyppeteer            0.0.25                 
-pyproj               2.5.0                  
-pyrsistent           0.16.0                 
-PySocks              1.7.1                  
-python-dateutil      2.8.1                  
-pytz                 2019.3                 
-pyviz-comms          0.7.4                  
-PyWavelets           1.1.1                  
-PyYAML               5.3.1                  
-pyzmq                19.0.0                 
-requests             2.23.0                 
-rmm                  0.13.0                 
-Rtree                0.9.4                  
-scikit-image         0.16.2                 
-scikit-learn         0.22.2.post1           
-scipy                1.4.1                  
-Send2Trash           1.5.0                  
-setuptools           46.1.3.post20200325    
-Shapely              1.7.0                  
-simpervisor          0.3                    
-six                  1.14.0                 
-sortedcontainers     2.1.0                  
-tblib                1.6.0                  
-terminado            0.8.3                  
-testpath             0.4.4                  
-toolz                0.10.0                 
-tornado              6.0.4                  
-tqdm                 4.45.0                 
-traitlets            4.3.3                  
-ucx-py               0.13.0                 
-urllib3              1.25.7                 
-wcwidth              0.1.9                  
-webencodings         0.5.1                  
-websockets           8.1                    
-wheel                0.34.2                 
-xarray               0.15.1                 
-xlrd                 1.2.0                  
-yarl                 1.3.0                  
-zict                 2.0.0                  
-zipp                 3.1.0
